backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
scikit-image
python-multipart
